Move aux entities to their own folder
