+++ core/entities/arc.js
@@ -10,8 +10,8 @@ import { AddState, RemoveState } from '../lib/stateManager.js';
 
 import { DesignCore } from '../designCore.js';
 import { Property } from '../properties/property.js';
-import { SnapPoint } from '../lib/snapPoint.js';
-import { RubberBand } from '../lib/rubberBand.js';
+import { SnapPoint } from '../lib/auxiliary/snapPoint.js';
+import { RubberBand } from '../lib/auxiliary/rubberBand.js';
 
 /**
  * Arc Entity Class
