+++ core/entities/circle.js
@@ -10,8 +10,8 @@ import { Utils } from '../lib/utils.js';
 
 import { DesignCore } from '../designCore.js';
 import { AddState, RemoveState } from '../lib/stateManager.js';
-import { SnapPoint } from '../lib/snapPoint.js';
-import { RubberBand } from '../lib/rubberBand.js';
+import { SnapPoint } from '../lib/auxiliary/snapPoint.js';
+import { RubberBand } from '../lib/auxiliary/rubberBand.js';
 
 /**
  * Circle Entity Class
