+++ core/entities/text.js
@@ -10,8 +10,8 @@ import { Flags } from '../properties/flags.js';
 import { Property } from '../properties/property.js';
 
 import { DesignCore } from '../designCore.js';
-import { SnapPoint } from '../lib/snapPoint.js';
-import { RubberBand } from '../lib/rubberBand.js';
+import { SnapPoint } from '../lib/auxiliary/snapPoint.js';
+import { RubberBand } from '../lib/auxiliary/rubberBand.js';
 
 /**
  * Text Entity Class
