+++ core/tools/rotate.js
@@ -7,7 +7,7 @@ import { Point } from '../entities/point.js';
 import { UpdateState } from '../lib/stateManager.js';
 
 import { DesignCore } from '../designCore.js';
-import { RubberBand } from '../lib/rubberBand.js';
+import { RubberBand } from '../lib/auxiliary/rubberBand.js';
 
 /**
  * Rotate Command Class
